automation_app/requirements.txt
pandas==2.2.2
allure-pytest==2.13.5
openpyxl==3.1.2
requests==2.31.0
jsonschema==4.21.1
pytest==8.1.1
pytest-order==1.0.1
pytest-xdist==3.5.0
python-dotenv==1.0.1
numpy==1.26.4

selenium==4.3.0
webdriver-manager
bs4
flask
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0